requirements.txt
dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
